+++ source/Validation/documents/release-notes/release-notes.md
@@ -1,5 +1,11 @@
 # Schema Validation Release notes
 
+## Version 1.0.3
+
+- Added CreateErrorResponse extension method.
+- Added WriteAsXmlAsync extension method.
+- Added AsXmlReaderAsync extension method.
+
 ## Version 1.0.2
 
 - Added AsXElementAsync extension method.